feat: adjust app tools dir structure
