+++ packages/solutions/app-tools/src/compat/utils.ts
