+++ packages/solutions/app-tools/src/constants.ts
