+++ packages/solutions/app-tools/src/utils/getConfigFile.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import { CONFIG_FILE_EXTENSIONS, findExists } from '@modern-js/utils';
-import { DEFAULT_CONFIG_FILE } from './constants';
+import { DEFAULT_CONFIG_FILE } from '../constants';
 
 export const getConfigFile = (configFile?: string) =>
   findExists(
