+++ packages/solutions/app-tools/src/utils/initAppContext.ts
