+++ packages/solutions/app-tools/src/utils/isAutoLoadPlugins.ts
@@ -1,5 +1,5 @@
 import { createLoadedConfig } from '@modern-js/plugin-v2/cli';
-export async function getIsAutoLoadPlugins(
+export async function isAutoLoadPlugins(
   appDirectory: string,
   configFile = 'modern.config.ts',
   packageJsonConfig = 'ModernConfig',
